add ability to use don't-care maps with source and target
The system image has don't-care regions where we can't count on what
data is present there.  (Some tools may write blocks of zeros but
others may not, and writing blocks of zeros interferes with the
wear-leveling mechanisms on flash.)  Modify syspatch to support a
"don't care map", which is basically a list describing the size of
care and don't-care regions in the partition, and use this when
reading the source and writing the target.

Certain pathological distributions of don't care blocks can make the
target-writing overlap the source-reading.  The current code detects
this and crashes; correct handling requires buffering the target data
until the source reading moves ahead.

Change-Id: I397b10494b17148f8bbf58a2ca98da57b288e1a5