Don't increment the window count for 0-length windows.
These windows should be consumed like any others, but should not
be added to the window count or the write queue.

This issue does not cause frontier violations per se, but does
contribute to the problem by shortening the effective length of
the write queue. This is particularly noticable in- but not unique
to- the situation where you have a read cache shorter than one
full source window.

Change-Id: Iba9739b9a82b60bcd7029652fe3bb6c4d2892eca