Increment TARGET_WINDOWS_WRITTEN after stream_to_target_write.
Prior to this, TARGET_WINDOWS_WRITTEN would be incremented before
the call to stream_to_target_write if the target was nonzero size
and write_target succeeded. This caused tgt->start to be set
incorrectly. This confused mapio but not xdelta, leading to
correct target windows written to incorrect locations.

Change-Id: I76db5281a749a20d0e63b8126238eb1886d87a1c